requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
pyyaml
